refactor: Change nested stacks into constructs
