+++ infrastructure/constructs/bundled_code.py
@@ -3,8 +3,8 @@ from os.path import join
 from aws_cdk import aws_lambda
 from aws_cdk.core import BundlingOptions
 
-from ..runtime import PYTHON_RUNTIME
 from .backend import BACKEND_DIRECTORY
+from .runtime import PYTHON_RUNTIME
 
 
 def bundled_code(directory: str) -> aws_lambda.Code:
