+++ infrastructure/constructs/bundled_lambda_function.py
@@ -3,10 +3,10 @@ from typing import Mapping, Optional
 from aws_cdk import aws_lambda, aws_lambda_python
 from aws_cdk.core import Construct, Duration
 
-from ..common import LOG_LEVEL
-from ..runtime import PYTHON_RUNTIME
 from .backend import BACKEND_DIRECTORY
 from .bundled_code import bundled_code
+from .common import LOG_LEVEL
+from .runtime import PYTHON_RUNTIME
 
 
 class BundledLambdaFunction(aws_lambda.Function):
