+++ infrastructure/constructs/lambda_endpoint.py
@@ -1,9 +1,9 @@
 from aws_cdk import aws_iam, aws_lambda, aws_lambda_python
 from aws_cdk.core import Construct, Duration
 
-from ..runtime import PYTHON_RUNTIME
 from .backend import BACKEND_DIRECTORY
 from .bundled_code import bundled_code
+from .runtime import PYTHON_RUNTIME
 
 
 class LambdaEndpoint(aws_lambda.Function):
