+++ infrastructure/constructs/removal_policy.py
